Remove MessagePackReaderExtensions, move tests to MsgPackReaderTests
